Refactor HashAggregationOperator
Extracted interface for HashAggregationBuilder. This change will make
it easier to implement future spillable version.
