+++ presto-main/src/main/java/com/facebook/presto/operator/HashAggregationOperator.java
@@ -15,29 +15,24 @@ package com.facebook.presto.operator;
 
 import com.facebook.presto.operator.aggregation.Accumulator;
 import com.facebook.presto.operator.aggregation.AccumulatorFactory;
-import com.facebook.presto.operator.aggregation.GroupedAccumulator;
+import com.facebook.presto.operator.aggregation.builder.HashAggregationBuilder;
+import com.facebook.presto.operator.aggregation.builder.InMemoryHashAggregationBuilder;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageBuilder;
-import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.sql.planner.plan.AggregationNode.Step;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.facebook.presto.type.TypeUtils;
-import com.google.common.collect.AbstractIterator;
 import com.google.common.collect.ImmutableList;
-import com.google.common.primitives.Ints;
 import io.airlift.units.DataSize;
 import io.airlift.units.DataSize.Unit;
 
-import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import static com.facebook.presto.operator.GroupByHash.createGroupByHash;
-import static com.facebook.presto.spi.type.BigintType.BIGINT;
-import static com.google.common.base.Preconditions.checkArgument;
+import static com.facebook.presto.operator.aggregation.builder.InMemoryHashAggregationBuilder.toTypes;
 import static com.google.common.base.Preconditions.checkState;
 import static java.util.Objects.requireNonNull;
 
@@ -59,8 +54,9 @@ public class HashAggregationOperator
 
         private final int expectedGroups;
         private final List<Type> types;
-        private boolean closed;
         private final long maxPartialMemory;
+
+        private boolean closed;
 
         public HashAggregationOperatorFactory(
                 int operatorId,
@@ -157,7 +153,7 @@ public class HashAggregationOperator
 
     private final List<Type> types;
 
-    private GroupByHashAggregationBuilder aggregationBuilder;
+    private HashAggregationBuilder aggregationBuilder;
     private Iterator<Page> outputIterator;
     private boolean inputProcessed;
     private boolean finishing;
@@ -228,7 +224,7 @@ public class HashAggregationOperator
         inputProcessed = true;
 
         if (aggregationBuilder == null) {
-            aggregationBuilder = new GroupByHashAggregationBuilder(
+            aggregationBuilder = new InMemoryHashAggregationBuilder(
                     accumulatorFactories,
                     step,
                     expectedGroups,
@@ -274,7 +270,7 @@ public class HashAggregationOperator
                 return null;
             }
 
-            outputIterator = aggregationBuilder.build();
+            outputIterator = aggregationBuilder.buildResult();
             aggregationBuilder = null;
 
             if (!outputIterator.hasNext()) {
@@ -321,186 +317,5 @@ public class HashAggregationOperator
             return null;
         }
         return output.build();
-    }
-
-    private static List<Type> toTypes(List<? extends Type> groupByType, Step step, List<AccumulatorFactory> factories, Optional<Integer> hashChannel)
-    {
-        ImmutableList.Builder<Type> types = ImmutableList.builder();
-        types.addAll(groupByType);
-        if (hashChannel.isPresent()) {
-            types.add(BIGINT);
-        }
-        for (AccumulatorFactory factory : factories) {
-            types.add(new Aggregator(factory, step).getType());
-        }
-        return types.build();
-    }
-
-    private static class GroupByHashAggregationBuilder
-    {
-        private final GroupByHash groupByHash;
-        private final List<Aggregator> aggregators;
-        private final OperatorContext operatorContext;
-        private final boolean partial;
-
-        private boolean full;
-
-        private GroupByHashAggregationBuilder(
-                List<AccumulatorFactory> accumulatorFactories,
-                Step step,
-                int expectedGroups,
-                List<Type> groupByTypes,
-                List<Integer> groupByChannels,
-                Optional<Integer> hashChannel,
-                OperatorContext operatorContext)
-        {
-            this.groupByHash = createGroupByHash(operatorContext.getSession(), groupByTypes, Ints.toArray(groupByChannels), hashChannel, expectedGroups);
-            this.operatorContext = operatorContext;
-            this.partial = step.isOutputPartial();
-
-            // wrapper each function with an aggregator
-            ImmutableList.Builder<Aggregator> builder = ImmutableList.builder();
-            requireNonNull(accumulatorFactories, "accumulatorFactories is null");
-            for (int i = 0; i < accumulatorFactories.size(); i++) {
-                AccumulatorFactory accumulatorFactory = accumulatorFactories.get(i);
-                builder.add(new Aggregator(accumulatorFactory, step));
-            }
-            aggregators = builder.build();
-        }
-
-        private void processPage(Page page)
-        {
-            if (aggregators.isEmpty()) {
-                groupByHash.addPage(page);
-                return;
-            }
-
-            GroupByIdBlock groupIds = groupByHash.getGroupIds(page);
-
-            for (Aggregator aggregator : aggregators) {
-                aggregator.processPage(groupIds, page);
-            }
-            updateMemory();
-        }
-
-        private void updateMemory()
-        {
-            long memorySize = groupByHash.getEstimatedSize();
-            for (Aggregator aggregator : aggregators) {
-                memorySize += aggregator.getEstimatedSize();
-            }
-            memorySize -= operatorContext.getOperatorPreAllocatedMemory().toBytes();
-            if (memorySize < 0) {
-                memorySize = 0;
-            }
-            if (partial) {
-                full = !operatorContext.trySetMemoryReservation(memorySize);
-            }
-            else {
-                operatorContext.setMemoryReservation(memorySize);
-            }
-        }
-
-        public boolean isFull()
-        {
-            return full;
-        }
-
-        public Iterator<Page> build()
-        {
-            List<Type> types = new ArrayList<>(groupByHash.getTypes());
-            for (Aggregator aggregator : aggregators) {
-                types.add(aggregator.getType());
-            }
-
-            final PageBuilder pageBuilder = new PageBuilder(types);
-            return new AbstractIterator<Page>()
-            {
-                private final int groupCount = groupByHash.getGroupCount();
-                private int groupId;
-
-                @Override
-                protected Page computeNext()
-                {
-                    if (groupId >= groupCount) {
-                        return endOfData();
-                    }
-
-                    pageBuilder.reset();
-
-                    List<Type> types = groupByHash.getTypes();
-                    while (!pageBuilder.isFull() && groupId < groupCount) {
-                        groupByHash.appendValuesTo(groupId, pageBuilder, 0);
-
-                        pageBuilder.declarePosition();
-                        for (int i = 0; i < aggregators.size(); i++) {
-                            Aggregator aggregator = aggregators.get(i);
-                            BlockBuilder output = pageBuilder.getBlockBuilder(types.size() + i);
-                            aggregator.evaluate(groupId, output);
-                        }
-
-                        groupId++;
-                    }
-
-                    return pageBuilder.build();
-                }
-            };
-        }
-    }
-
-    private static class Aggregator
-    {
-        private final GroupedAccumulator aggregation;
-        private final Step step;
-        private final int intermediateChannel;
-
-        private Aggregator(AccumulatorFactory accumulatorFactory, Step step)
-        {
-            if (step.isInputRaw()) {
-                intermediateChannel = -1;
-                aggregation = accumulatorFactory.createGroupedAccumulator();
-            }
-            else {
-                checkArgument(accumulatorFactory.getInputChannels().size() == 1, "expected 1 input channel for intermediate aggregation");
-                intermediateChannel = accumulatorFactory.getInputChannels().get(0);
-                aggregation = accumulatorFactory.createGroupedIntermediateAccumulator();
-            }
-            this.step = step;
-        }
-
-        public long getEstimatedSize()
-        {
-            return aggregation.getEstimatedSize();
-        }
-
-        public Type getType()
-        {
-            if (step.isOutputPartial()) {
-                return aggregation.getIntermediateType();
-            }
-            else {
-                return aggregation.getFinalType();
-            }
-        }
-
-        public void processPage(GroupByIdBlock groupIds, Page page)
-        {
-            if (step.isInputRaw()) {
-                aggregation.addInput(groupIds, page);
-            }
-            else {
-                aggregation.addIntermediate(groupIds, page.getBlock(intermediateChannel));
-            }
-        }
-
-        public void evaluate(int groupId, BlockBuilder output)
-        {
-            if (step.isOutputPartial()) {
-                aggregation.evaluateIntermediate(groupId, output);
-            }
-            else {
-                aggregation.evaluateFinal(groupId, output);
-            }
-        }
     }
 }
